refactor package structure
