+++ packages/remark-graphviz/package.json
@@ -1,13 +1,8 @@
 {
     "name": "@tdev/remark-graphviz",
     "version": "1.0.0",
-    "main": "lib/index.js",
-    "types": "lib/remark-pdf.d.ts",
-    "scripts": {
-        "build": "tsc",
-        "test": "vitest --coverage",
-        "dev": "tsc --watch"
-    },
+    "main": "remark-plugin/index.ts",
+    "types": "remark-plugin/index.ts",
     "dependencies": {
         "@hpcc-js/wasm-graphviz": "^1.7.0"
     },
