+++ packages/remark-pdf/package.json
@@ -1,8 +1,8 @@
 {
     "name": "@tdev/remark-pdf",
     "version": "1.0.0",
-    "main": "src/index.ts",
-    "types": "src/index.ts",
+    "main": "index.ts",
+    "types": "index.ts",
     "dependencies": {
         "react-pdf": "^9.2.1"
     },
