+++ packages/remark-pdf/remark-plugin/tests/assets/stairs.pdf
